# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loyalty & Referral System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CustomerMembershipLevel = typeof CustomerMembershipLevel[keyof typeof CustomerMembershipLevel];


export const CustomerMembershipLevel = {
  Bronze: 'Bronze',
  Silver: 'Silver',
  Gold: 'Gold',
  Platinum: 'Platinum',
} as const;

export interface Customer {
  id: number;
  /** @nullable */
  clerkUserId?: string | null;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  referralCode: string;
  /** @nullable */
  referredByCode?: string | null;
  totalPoints: number;
  totalSpent: number;
  servicesCount: number;
  referralCount: number;
  membershipLevel: CustomerMembershipLevel;
  createdAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone?: string;
  referredByCode?: string;
}

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  clerkUserId?: string;
}

export interface Service {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  pointsEarned: number;
  createdAt: string;
}

export interface ServiceInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  amount: number;
  /** @nullable */
  customPoints?: number | null;
}

export type PointTransactionType = typeof PointTransactionType[keyof typeof PointTransactionType];


export const PointTransactionType = {
  earn: 'earn',
  deduct: 'deduct',
  referral_bonus: 'referral_bonus',
} as const;

export interface PointTransaction {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  points: number;
  type: PointTransactionType;
  reason: string;
  /** @nullable */
  serviceId?: number | null;
  /** @nullable */
  referralId?: number | null;
  createdAt: string;
}

export type PointAdjustmentType = typeof PointAdjustmentType[keyof typeof PointAdjustmentType];


export const PointAdjustmentType = {
  earn: 'earn',
  deduct: 'deduct',
} as const;

export interface PointAdjustment {
  /** @minimum 1 */
  points: number;
  type: PointAdjustmentType;
  /** @minLength 1 */
  reason: string;
}

export type ReferralStatus = typeof ReferralStatus[keyof typeof ReferralStatus];


export const ReferralStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
} as const;

export type ReferralRewardType = typeof ReferralRewardType[keyof typeof ReferralRewardType];


export const ReferralRewardType = {
  points: 'points',
  percentage: 'percentage',
} as const;

export interface Referral {
  id: number;
  referrerId: number;
  /** @nullable */
  referrerName?: string | null;
  referredId: number;
  /** @nullable */
  referredName?: string | null;
  status: ReferralStatus;
  rewardType: ReferralRewardType;
  rewardValue: number;
  createdAt: string;
}

export type ReferralUpdateStatus = typeof ReferralUpdateStatus[keyof typeof ReferralUpdateStatus];


export const ReferralUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
} as const;

export interface ReferralUpdate {
  status?: ReferralUpdateStatus;
}

export type CommissionStatus = typeof CommissionStatus[keyof typeof CommissionStatus];


export const CommissionStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface Commission {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  referralId: number;
  amount: number;
  status: CommissionStatus;
  createdAt: string;
}

export type CommissionUpdateStatus = typeof CommissionUpdateStatus[keyof typeof CommissionUpdateStatus];


export const CommissionUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface CommissionUpdate {
  status?: CommissionUpdateStatus;
}

export interface DashboardStats {
  totalCustomers: number;
  totalPoints: number;
  totalRevenue: number;
  totalReferrals: number;
  totalCommissionsPaid: number;
  pendingCommissions: number;
}

export type MembershipCountLevel = typeof MembershipCountLevel[keyof typeof MembershipCountLevel];


export const MembershipCountLevel = {
  Bronze: 'Bronze',
  Silver: 'Silver',
  Gold: 'Gold',
  Platinum: 'Platinum',
} as const;

export interface MembershipCount {
  level: MembershipCountLevel;
  count: number;
}

export interface DiscountTier {
  pointsRequired: number;
  discountPercent: number;
  achieved: boolean;
}

export interface DiscountInfo {
  currentPoints: number;
  availableDiscount: number;
  discountTiers: DiscountTier[];
}