# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loyalty & Referral System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /admin/customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ── Admin: Services ─────────────────────────────────────────────────────────
  /admin/customers/{customerId}/services:
    post:
      operationId: addService
      tags: [services]
      summary: Add a service to a customer (admin)
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /admin/customers/{customerId}/points:
    post:
      operationId: adjustPoints
      tags: [points]
      summary: Add or deduct points for a customer (admin)
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointAdjustment"
      responses:
        "201":
          description: Point transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointTransaction"
...
  /admin/referrals/{id}:
    patch:
      operationId: updateReferral
      tags: [referrals]
      summary: Update referral status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"

  # ── Admin: Commissions ──────────────────────────────────────────────────────
...
  /admin/commissions/{id}:
    patch:
      operationId: updateCommission
      tags: [commissions]
      summary: Update commission status (paid/unpaid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"

  # ── Admin: Dashboard ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, email, referralCode, totalPoints, totalSpent, servicesCount, referralCount, membershipLevel, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        referralCode:
          type: string
        referredByCode:
          type: ["string", "null"]
        totalPoints:
          type: integer
        totalSpent:
          type: number
        servicesCount:
          type: integer
        referralCount:
          type: integer
        membershipLevel:
          type: string
          enum: [Bronze, Silver, Gold, Platinum]
        createdAt:
          type: string
          format: date-time
    CustomerInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        referredByCode:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        clerkUserId:
          type: string
    Service:
      type: object
      required: [id, customerId, name, amount, pointsEarned, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        pointsEarned:
          type: integer
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required: [name, amount]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        amount:
          type: number
          minimum: 0
        customPoints:
          type: ["integer", "null"]
    PointTransaction:
      type: object
      required: [id, customerId, points, type, reason, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        points:
          type: integer
        type:
          type: string
          enum: [earn, deduct, referral_bonus]
        reason:
          type: string
        serviceId:
          type: ["integer", "null"]
        referralId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    PointAdjustment:
      type: object
      required: [points, type, reason]
      properties:
        points:
          type: integer
          minimum: 1
        type:
          type: string
          enum: [earn, deduct]
        reason:
          type: string
          minLength: 1
    Referral:
      type: object
      required: [id, referrerId, referredId, status, rewardType, rewardValue, createdAt]
      properties:
        id:
          type: integer
        referrerId:
          type: integer
        referrerName:
          type: ["string", "null"]
        referredId:
          type: integer
        referredName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, rejected]
        rewardType:
          type: string
          enum: [points, percentage]
        rewardValue:
          type: number
        createdAt:
          type: string
          format: date-time
    ReferralUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, rejected]
    Commission:
      type: object
      required: [id, customerId, referralId, amount, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        referralId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid]
        createdAt:
          type: string
          format: date-time
    CommissionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid]
...