# lib/api-zod/src/generated/types/commissionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loyalty & Referral System API
 * OpenAPI spec version: 0.1.0
 */
import type { CommissionUpdateStatus } from './commissionUpdateStatus';

export interface CommissionUpdate {
  status?: CommissionUpdateStatus;
}