# lib/api-zod/src/generated/types/pointTransaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loyalty & Referral System API
 * OpenAPI spec version: 0.1.0
 */
import type { PointTransactionType } from './pointTransactionType';

export interface PointTransaction {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  points: number;
  type: PointTransactionType;
  reason: string;
  /** @nullable */
  serviceId?: number | null;
  /** @nullable */
  referralId?: number | null;
  createdAt: Date;
}