# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loyalty & Referral System API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  pointsEarned: number;
  createdAt: Date;
}